refactor(#138): controller 분리
- websocket과 rest controller 구분을 하였습니다.
- swagger를 작성하였습니다.
